moved FIG blocks into their own discrete file
